+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.Dto.WinningLottoDto;
+import lotto.dto.WinningLottoDto;
 import lotto.domain.TicketModel.*;
 
 public class WinningLotto extends AbstractWinningLotto {
